feat: lista de cambios aplicados:
### 1. Generación de IDs no thread-safe en Transaction.java

### 2. Imports duplicados en TransactionRepository.java

### 3. Referencia a constante no declarada en TransactionRepository.java

### 4. SQL Injection potencial en Dialog.java

### 5. Race condition en getSaldosPorCuenta - TransactionRepository.java

### 6. Uso de Double para dinero - múltiples archivos

### 7. Falta de cierre de recursos en WalletRepository.java

### 8. NotifyDataSetChanged() ineficiente en TransactionAdapter.java

### 9. Validación de entrada insuficiente en AgregarFragment.java

### 10. Memory leak potencial en Dialog.java

### 11. Comentarios en español mezclados con código

### 12. Hardcoded strings en múltiples archivos

### 13. Falta de null safety en TransactionMapper.java

### 14. Calendar.getInstance() llamado repetidamente

### 15. Falta de manejo de excepciones específico

### 16. ViewHolder con campos públicos

### 17. ExecutorService no cerrado

### 18. Falta de documentación en interfaces

### 19. Falta de equals() y hashCode() en modelos

### 20. NullPointerException riesgo en StringParser.java

### 21. Patrón lookup ineficiente en TransactionAdapter.java

### 22. Thread safety insuficiente en MyApplication.java

### 23. Falta de null check en getAppName() - Wallets.java

+++ app/src/main/java/com/notificationcapture/app/database/CreditCardPaymentDao.java
@@ -6,6 +6,8 @@ import androidx.room.OnConflictStrategy;
 import androidx.room.Query;
 import androidx.lifecycle.LiveData;
 
+import java.math.BigDecimal;
+
 @Dao
 public interface CreditCardPaymentDao {
 
@@ -15,20 +17,20 @@ public interface CreditCardPaymentDao {
     @Query("SELECT COALESCE(SUM(montoPagado), 0) FROM credit_card_payments " +
             "WHERE creditCardId = :creditCardId " +
             "AND startTimestamp >= :start - 1000 AND startTimestamp <= :start + 1000")
-    double getTotalPagadoEnPeriodo(String creditCardId, long start);
+    BigDecimal getTotalPagadoEnPeriodo(String creditCardId, long start);
 
     @Query("SELECT COALESCE(SUM(montoPagado), 0) FROM credit_card_payments " +
             "WHERE creditCardId = :creditCardId " +
             "AND startTimestamp < :start")
-    double getPagosAnteriores(String creditCardId, long start);
+    BigDecimal getPagosAnteriores(String creditCardId, long start);
 
     @Query("SELECT COALESCE(SUM(montoPagado), 0) FROM credit_card_payments " +
             "WHERE creditCardId = :creditCardId " +
             "AND startTimestamp >= :start - 1000 AND startTimestamp <= :start + 1000")
-    LiveData<Double> getTotalPagadoEnPeriodoLiveData(String creditCardId, long start);
+    LiveData<BigDecimal> getTotalPagadoEnPeriodoLiveData(String creditCardId, long start);
 
     @Query("SELECT COALESCE(SUM(montoPagado), 0) FROM credit_card_payments " +
             "WHERE creditCardId = :creditCardId " +
             "AND startTimestamp < :start")
-    LiveData<Double> getPagosAnterioresLiveData(String creditCardId, long start);
+    LiveData<BigDecimal> getPagosAnterioresLiveData(String creditCardId, long start);
 }
