+++ app/src/main/java/com/notificationcapture/app/interfaces/OnCategoryClickListener.java
@@ -1,5 +1,7 @@
 package com.notificationcapture.app.interfaces;
 
+import java.math.BigDecimal;
+
 public interface OnCategoryClickListener {
-    void onCategoryClick(String category, Double totalAmount);
+    void onCategoryClick(String category, BigDecimal totalAmount);
 }
