+++ app/src/main/java/com/notificationcapture/app/interfaces/TransferListener.java
@@ -1,5 +1,7 @@
 package com.notificationcapture.app.interfaces;
 
+import java.math.BigDecimal;
+
 public interface TransferListener {
-    void onTransferConfirmed(String origenId, String destinoId, double monto);
+    void onTransferConfirmed(String origenId, String destinoId, BigDecimal monto);
 }
